feat: add new pages, refactor routing and add search router
