+++ index.js
@@ -3,6 +3,9 @@ const env = process.env.NODE_ENV || 'development'
 const config = require('./config/config')[env]
 const express = require('express')
 const indexRouter = require('./routes')
+const authRouter = require('./routes/auth')
+const cubeRouter = require('./routes/cube')
+const accessoryRouter = require('./routes/accessory')
 const app = express()
 
 const mongoose = require('mongoose')
@@ -21,6 +24,9 @@ mongoose.connect(config.dataBaseUrl, {
 
 require('./config/express')(app)
 
+app.use('/', authRouter)
+app.use('/', cubeRouter)
+app.use('/', accessoryRouter)
 app.use('/', indexRouter)
 
 app.listen(config.port, console.log(`Listening on port ${config.port}! Now its up to you...`))
