+++ views/index.hbs
@@ -1,6 +1,6 @@
 <main>
   <h1>Browser</h1>
-  <form action="" method="">
+  <form action="/search" method="POST">
     <input type="text" class="search" name="search" placeholder="Search...">
     <input type="number" name="from" class="difficulty" placeholder="Difficulty from...">
     <span>-</span>
