+++ views/partials/header.hbs
@@ -4,6 +4,9 @@
         <li><a href="/">Browse</a></li>
         <li><a href="/create">Add a Cube</a></li>
         <li><a href="/create/accessory">Create Accesory</a></li>
+        <li><a href="/login">Login</a></li>
+        <li><a href="/">Log Out</a></li>
+        <li><a href="/register">Register</a></li>
         <li><a href="/about">About</a></li>
     </ul>
 </header>
